requirements.txt
black==22.1.0
fastapi==0.73.0
isort==5.10.1
mypy==0.931
Pillow==9.0.1
python-multipart==0.0.5
qrcode==7.3.1
uvicorn==0.17.4
